+++ autotorrent/cmd.py
@@ -1,6 +1,7 @@
 import argparse
 import logging
 import os
+import shutil
 
 from six.moves import configparser
 
@@ -10,6 +11,7 @@ from autotorrent.db import Database
 def commandline_handler():
     parser = argparse.ArgumentParser()
     parser.add_argument("-c", "--config", dest="config_file", default="autotorrent.conf", help="Path to config file")
+    parser.add_argument("--create_config", dest="create_config_file", default="autotorrent.conf", help="Creates a new configuration file")
     
     parser.add_argument("-t", "--test_connection", action="store_true", dest="test_connection", default=False, help='Tests the connection to the torrent client')
     parser.add_argument("-r", "--rebuild", dest="rebuild", default=False, help='Rebuild the database', nargs='*')
@@ -20,6 +22,15 @@ def commandline_handler():
     args = parser.parse_args()
     
     logging.basicConfig(level=logging.DEBUG if args.verbose else logging.ERROR)
+    
+    if args.create_config_file:
+        if os.path.exists(args.create_config_file):
+            parser.error("Target %r already exists, not creating" % args.create_config_file)
+        else:
+            src = os.path.join(os.path.dirname(__file__), 'autotorrent.conf.dist')
+            shutil.copy(src, args.create_config_file)
+            print('Created configuration file %r' % args.create_config_file)
+        quit()
     
     if not os.path.isfile(args.config_file):
         parser.error("Config file not found %r" % args.config_file)